refactor: remove unused lines and move file until it feels right
