+++ src/details.html
@@ -47,7 +47,7 @@
       </div>
 
       <div id="basic-info"></div>
-      <div id="ability-container" class="my-8 flex flex-col items-center"></div>
+      <div id="ability-wrapper" class="my-8 flex flex-col items-center"></div>
       <div id="stats-wrapper"></div>
     </div>
   </body>
