+++ src/index.html
@@ -8,7 +8,7 @@
   </head>
   <body class="mx-auto flex max-w-screen-lg flex-col gap-6 p-4 md:p-10">
     <div class="navbar flex-col justify-between md:flex-row">
-      <a href="http://localhost:8080/">
+      <a href="./index.html">
         <img src="./assets/pokehunt.png" alt="pokehunt logo" class="h-16" />
       </a>
