+++ src/script/details.js
@@ -3,23 +3,27 @@ import "./components/basic-info-card";
 import "./components/stat-list";
 
 import { SCENERY } from "./constant/scenery";
-import { PokemonService } from "./pokemon.service";
+import { PokemonService } from "./utils/pokemon.service";
 import { snakeToTitleCase } from "./utils/string";
 
 const pokemonService = new PokemonService();
 
 const basicInfoWrapper = document.getElementById("basic-info");
-const abilityWrapper = document.getElementById("ability-container");
+const abilityWrapper = document.getElementById("ability-wrapper");
 const statsWrapper = document.getElementById("stats-wrapper");
 
-const renderHeader = ({ name, details, sprites }) => {
+const renderHeader = ({ name, details, sprites, id }) => {
   document
     .getElementById("scenery")
     .setAttribute("src", SCENERY[Math.floor(Math.random() * 4) + 1]);
   document.querySelector("h1").innerHTML = snakeToTitleCase(name);
 
   const spriteImg = document.getElementById("pokemon-sprite");
-  spriteImg.setAttribute("src", sprites.animated);
+  spriteImg.setAttribute(
+    "src",
+    sprites.animated ??
+      `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${id}.png`,
+  );
   spriteImg.setAttribute("alt", `${name} image`);
 
   const infoCard = document.createElement("basic-info-card");
@@ -44,7 +48,7 @@ const details = () => {
   const data = pokemonService.pokemonDetails;
 
   if (!queryParams.has("id") || data.id === undefined) {
-    window.location.href = "http://localhost:8080";
+    window.location.href = `${window.location.protocol}//${window.location.host}`;
   }
 
   renderHeader(data);
